Add baseline gitignore
A gitignore for the project didn't exist until now. While this
doesn't have to do directly with the work of the current PR, it seems
good to me to just have _something_ in place for future gitignores, and
no time like the present.

Throwing a .DS_Store ignore in there to just be extra cautious that
those pesky OS X/macOS files don't make their way into the repo.