Add custom fake resolver
This adds a custom resolver, so that when scanning the IP address
we scan for a given domain name is the one we specify, not whatever
the http libraries resolves for us.

So that redirects continue to work as expected, latch the lookups
beyond the original hostname, and call a shadow dialer that will
talk to a real DNS server.