Introduce ScanModuleAttributes
Rather than add a bunch of accessor functions that return static values
for each module, abstract the into ScanModuleAttributes, and add a
single function that exposes this. Previously, much of this information
was exposed as part of the call to RegisterModule. That remains there
for backwards compatibility (for now), but this allows it to be exposed
programatically, and continues work towards removing global state.