Use bodyTextLen instead of readLen for FailHTTPToHTTPS logic
respContentLength can be -1 in certain cases, in which case
readLen will be maxReadLen for the current scan. This will,
however, then cause the FailHTTPToHTTPS if-condition to
fail as the readLen is > 1024, even though the body content
length can be in this range. So let us use the actual body
length for the check to avoid this issue.