Deal with non-RFC compliant servers
There are hundreds of thousands of servers on the internet where the
"INTENT" is to be an HTTP service, but in reality they do not properly
conform to the HTTP RFC. In many cases, these servers will send back a
simple:

HTTP/1.1 404 Not Found

With no `Connection: close` header, and terminates the established
connection.

While it's understood these are not RFC compliant, I would still like to
see the servers that are intending to be HTTP servers.